+++ packages/perseus-editor/src/components/__stories__/locked-point-settings.stories.tsx
@@ -6,7 +6,7 @@ import {getDefaultFigureForType} from "../util";
 import type {Meta, StoryObj} from "@storybook/react";
 
 export default {
-    title: "Perseus Editor/Components/Locked Point Settings",
+    title: "PerseusEditor/Components/Locked Point Settings",
     component: LockedPointSettings,
 } as Meta<typeof LockedPointSettings>;
 
@@ -20,13 +20,15 @@ type StoryComponentType = StoryObj<typeof LockedPointSettings>;
 Default.args = {
     ...getDefaultFigureForType("point"),
     onChangeProps: () => {},
+    onRemove: () => {},
 };
 
 export const Controlled: StoryComponentType = {
     render: function Render() {
-        const [props, setProps] = React.useState(
-            getDefaultFigureForType("point"),
-        );
+        const [props, setProps] = React.useState({
+            ...getDefaultFigureForType("point"),
+            onRemove: () => {},
+        });
 
         const handlePropsUpdate = (newProps) => {
             setProps({