+++ packages/perseus/src/widgets/__tests__/interactive-graph.test.ts
@@ -19,6 +19,12 @@ import {
     polygonQuestionDefaultCorrect,
     pointQuestionWithDefaultCorrect,
     segmentWithLockedLineQuestion,
+    segmentQuestion,
+    linearQuestion,
+    linearSystemQuestion,
+    rayQuestion,
+    polygonQuestion,
+    pointQuestion,
 } from "../__testdata__/interactive-graph.testdata";
 import {trueForAllMafsSupportedGraphTypes} from "../interactive-graphs/mafs-supported-graph-types";
 
@@ -139,6 +145,17 @@ describe("mafs graphs", () => {
     const graphQuestionRenderers: {
         [K in (typeof mafsSupportedGraphTypes)[number]]: PerseusRenderer;
     } = {
+        segment: segmentQuestion,
+        linear: linearQuestion,
+        "linear-system": linearSystemQuestion,
+        ray: rayQuestion,
+        polygon: polygonQuestion,
+        point: pointQuestion,
+    };
+
+    const graphQuestionRenderersCorrect: {
+        [K in (typeof mafsSupportedGraphTypes)[number]]: PerseusRenderer;
+    } = {
         segment: segmentQuestionDefaultCorrect,
         linear: linearQuestionWithDefaultCorrect,
         "linear-system": linearSystemQuestionWithDefaultCorrect,
@@ -163,6 +180,15 @@ describe("mafs graphs", () => {
 
                 // Assert
                 expect(renderer).toHaveInvalidInput();
+            });
+        },
+    );
+
+    describe.each(Object.entries(graphQuestionRenderersCorrect))(
+        "graph type %s: default correct",
+        (_type, question) => {
+            it("should render", () => {
+                renderQuestion(question, apiOptions);
             });
 
             it("rejects incorrect answer", async () => {
@@ -342,9 +368,12 @@ describe("locked layer", () => {
         // Act
         // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
         const lines = container.querySelectorAll(".locked-line");
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        const rays = container.querySelectorAll(".locked-ray");
 
         // Assert
-        expect(lines).toHaveLength(3);
+        expect(lines).toHaveLength(2);
+        expect(rays).toHaveLength(1);
     });
 
     test("should render locked lines with styles", () => {
@@ -361,14 +390,12 @@ describe("locked layer", () => {
         // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
         const lines = container.querySelectorAll(".locked-line line");
         // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
-        const ray = container.querySelector(".locked-line g");
+        const ray = container.querySelector(".locked-ray g");
 
         // Assert
         expect(lines).toHaveLength(3);
         expect(lines[0]).toHaveStyle({stroke: lockedFigureColors.purple});
         expect(lines[1]).toHaveStyle({stroke: lockedFigureColors.green});
-        // Rays are handled a little differently - their color is applied
-        // to the g rather than to the line element.
         expect(ray).toHaveStyle({stroke: lockedFigureColors.pink});
     });
 
@@ -384,28 +411,30 @@ describe("locked layer", () => {
 
         // Act
         // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
-        const points = container.querySelectorAll(".locked-line circle");
+        const linePoints = container.querySelectorAll(".locked-line circle");
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        const rayPoints = container.querySelectorAll(".locked-ray circle");
 
         // Assert
-        expect(points).toHaveLength(5);
+        expect(linePoints).toHaveLength(4);
         // Two points for each line
-        expect(points[0]).toHaveStyle({
+        expect(linePoints[0]).toHaveStyle({
             fill: lockedFigureColors.purple,
             stroke: lockedFigureColors.purple,
         });
-        expect(points[1]).toHaveStyle({
+        expect(linePoints[1]).toHaveStyle({
             fill: wbColor.white,
             stroke: lockedFigureColors.purple,
         });
-        expect(points[2]).toHaveStyle({
+        expect(linePoints[2]).toHaveStyle({
             fill: wbColor.white,
             stroke: lockedFigureColors.green,
         });
-        expect(points[3]).toHaveStyle({
+        expect(linePoints[3]).toHaveStyle({
             fill: lockedFigureColors.green,
             stroke: lockedFigureColors.green,
         });
-        expect(points[4]).toHaveStyle({
+        expect(rayPoints[0]).toHaveStyle({
             fill: wbColor.white,
             stroke: lockedFigureColors.pink,
         });