+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/vector.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import {useTransform} from "../use-transform";
+import {useTransformVectorsToPixels} from "../use-transform";
 
 import {SVGLine} from "./svg-line";
 
@@ -15,7 +15,8 @@ type Props = {
 
 export function Vector(props: Props) {
     const {tail, tip, color, style} = props;
-    const [tailPx, tipPx] = useTransform(tail, tip);
+    const [tailPx, tipPx] = useTransformVectorsToPixels(tail, tip);
+
     return (
         <g style={{stroke: color, strokeWidth: 2}}>
             <SVGLine start={tailPx} end={tipPx} style={style} />